added decimalToHexAdrress function and FormatDateFromUnix to formatAddress file
